Reworked models, began building auth

+++ server/src/main/java/com/spellrack/server/repository/DeckRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource(collectionResourceRel = "deck", path = "deck")
-public interface DeckRepository extends JpaRepository<Deck, Integer> {
-
+public interface DeckRepository extends JpaRepository<Deck, Long> {
+    // Deck findById(Long deckId);
 }
